# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/messages/sent": {
            "get": {
                "description": "Returns a list of messages with status `sent`, ordered by `sent_time` ascending.\nSupports cursor-based pagination via the `after` cursor.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "List sent messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Return messages sent after this timestamp (RFC3339)",
                        "name": "after",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of messages to return (1–50), defaults to 42",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.SentMessagesResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/scheduler/toggle": {
            "post": {
                "description": "Starts the scheduler if it is stopped, or stops it if it is running.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scheduler"
                ],
                "summary": "Toggle the message scheduler",
                "responses": {
                    "200": {
                        "description": "Scheduler started or stopped",
                        "schema": {
                            "$ref": "#/definitions/handler.StatusResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handler.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "model.Message": {
            "type": "object",
            "properties": {
                "attempt_count": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string"
                },
                "sent_time": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.MessageStatus"
                }
            }
        },
        "model.MessageStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "SENT",
                "FAILED"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusSent",
                "StatusFailed"
            ]
        },
        "service.SentMessagesResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Message"
                    }
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        }
    }
}